+++ dkg/pedersen/mod_test.go
@@ -138,9 +138,9 @@ func TestPedersen_Scenario(t *testing.T) {
 	addrs := make([]mino.Address, n)
 
 	for i := 0; i < n; i++ {
+		addr := minogrpc.ParseAddress("127.0.0.1", 0)
 
-		port := uint16(2000 + i)
-		minogrpc, err := minogrpc.NewMinogrpc("127.0.0.1", port, tree.NewRouter(minogrpc.AddressFactory{}))
+		minogrpc, err := minogrpc.NewMinogrpc(addr, tree.NewRouter(minogrpc.AddressFactory{}))
 		require.NoError(t, err)
 
 		defer minogrpc.GracefulStop()
